Change how elasticsearch connector returns data
For the elasticsearch connector search by id method, if a string input is given, then the return value should be a single dictionary representing a single es document. If an iterable is given, then the return value should be a dictionary mapping of es document ids to their document values.